refactor: extract script history click handling into a separate function
